+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.TouCheese"
         tools:targetApi="31">
         <service
-            android:name="com.example.presentation.main.FCMService"
+            android:name="com.example.presentation.main.view.FCMService"
             android:permission="com.google.android.c2dm.permission.SEND"
             android:exported="false">
             <intent-filter>
