+++ presentation/src/main/java/com/example/presentation/main/view/MainActivity.kt
@@ -1,11 +1,7 @@
 package com.example.presentation.main.view
 
 import android.Manifest
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.app.PendingIntent
 import android.content.Context
-import android.content.Intent
 import android.content.pm.PackageManager
 import android.os.Build
 import android.os.Bundle
@@ -13,20 +9,23 @@ import android.os.LocaleList
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.app.NotificationCompat
 import androidx.core.content.ContextCompat
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.setupWithNavController
 import com.example.presentation.R
 import com.example.presentation.databinding.ActivityMainBinding
 import com.example.presentation.main.vm.FcmViewModel
+import com.example.presentation.util.NotificationHelper
+import com.example.presentation.util.NotificationHelper.sendNotification
+import com.google.firebase.messaging.FirebaseMessagingService
+import com.google.firebase.messaging.RemoteMessage
 import dagger.hilt.android.AndroidEntryPoint
 import java.util.Locale
 
 @AndroidEntryPoint
 class MainActivity : AppCompatActivity() {
     private val viewModel: FcmViewModel by viewModels()
-    private val binding : ActivityMainBinding by lazy {
+    private val binding: ActivityMainBinding by lazy {
         ActivityMainBinding.inflate(layoutInflater)
     }
 
@@ -42,7 +41,8 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(binding.root)
 
-        val navHostFragment = supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
+        val navHostFragment =
+            supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
         val navController = navHostFragment.navController
 
         val bottomNavigationView = binding.bottomNavigation
@@ -56,14 +56,18 @@ class MainActivity : AppCompatActivity() {
 
     private fun askNotificationPermission() {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-            if (ContextCompat.checkSelfPermission(this, Manifest.permission.POST_NOTIFICATIONS) ==
-                PackageManager.PERMISSION_GRANTED
-            ) {
+            when {
+                ContextCompat.checkSelfPermission(this, Manifest.permission.POST_NOTIFICATIONS) ==
+                        PackageManager.PERMISSION_GRANTED -> {
+                }
 
-            } else if (shouldShowRequestPermissionRationale(Manifest.permission.POST_NOTIFICATIONS)) {
+                shouldShowRequestPermissionRationale(Manifest.permission.POST_NOTIFICATIONS) -> {
+                    NotificationHelper.showPermissionDeniedDialog(this)
+                }
 
-            } else {
-                requestPermissionLauncher.launch(Manifest.permission.POST_NOTIFICATIONS)
+                else -> {
+                    requestPermissionLauncher.launch(Manifest.permission.POST_NOTIFICATIONS)
+                }
             }
         }
     }
@@ -78,39 +82,16 @@ class MainActivity : AppCompatActivity() {
     }
 }
 
+class FCMService : FirebaseMessagingService(){
+    override fun onNewToken(token: String) {
+        super.onNewToken(token)
+    }
 
-object NotificationHelper {
-    fun sendNotification(context: Context, title: String, messageBody: String) {
-        val channelId = "fcm_default_channel"
-        val notificationId = System.currentTimeMillis().toInt()
-
-        val appContext = context.applicationContext
-
-        val intent = Intent(appContext, MainActivity::class.java)
-        val pendingIntent = PendingIntent.getActivity(
-            appContext, 0, intent, PendingIntent.FLAG_IMMUTABLE
-        )
-
-        val notificationBuilder = NotificationCompat.Builder(appContext, channelId)
-            .setSmallIcon(R.drawable.icon_favorite_24px)
-            .setContentTitle(title)
-            .setContentText(messageBody)
-            .setPriority(NotificationCompat.PRIORITY_HIGH)
-            .setAutoCancel(false)
-            .setContentIntent(pendingIntent)  // 알림 클릭 시 지정된 Intent 실행
-
-        val notificationManager =
-            appContext.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-
-        // 채널 설정
-        val channel = NotificationChannel(
-            channelId,
-            "Default Channel",
-            NotificationManager.IMPORTANCE_HIGH // 채널 중요도
-        )
-
-        notificationManager.createNotificationChannel(channel)
-
-        notificationManager.notify(notificationId, notificationBuilder.build())
+    override fun onMessageReceived(remoteMessage: RemoteMessage) {
+        remoteMessage.notification?.let {
+            val title = it.title ?: "Default Title"
+            val message = it.body ?: "Default Body"
+            sendNotification(applicationContext, title, message)
+        }
     }
 }
